+++ boosty_downloader/src/infrastructure/path_sanitizer.py
